moving code to component-login (the client side)
